test_delay bug fix

+++ acrn-unit-test/guest/x86/64/interrupt_fn.c
@@ -377,7 +377,7 @@ void test_delay(int time)
 {
 	__unused int count = 0;
 	u64 tsc;
-	tsc = rdtsc() + time*1000000000;
+	tsc = rdtsc() + ((u64)time *1000000000);
 
 	while (rdtsc() < tsc) {
 		;
